Boost test coverage
The patch attached increases the cover which YAML-Tiny's test suite provides to
its source code.

At the start of my work on YAML-Tiny, coverage, as measured by Devel::Cover,
was:

    file                    stmt    bran    cond     sub
    blib/lib/YAML/Tiny.pm   88.6    77.7    73.0    92.6

At the point represented by the patch, coverage has increased to:

    file                    stmt    bran    cond     sub
    blib/lib/YAML/Tiny.pm   91.3    83.0    73.0   100.0

There are areas in the code where, if it were cleaned up, coverage could
probably be improved further.  But at the very least, we now have coverage of
all subroutines in lib/YAML/Tiny.pm.