Multifile runner is a Util function.
testml_run_all_files doesn't belong in TestMLTiny even if if has
'TestML' in the name. It is a helper/utility function for running many
of the YAML::Tiny tests that happen to be written in TestML.

Move this function to TestUtils and rename it slightly to
run_all_testml_files.