Finish pruning and documenting. 53 tests.
xdg wondered why N/Y/n/y (and many other variants) were removed, so I'll
explain here:

The short answer is to match YAML::XS and (hopefully) ::Syck and .pm
(need to test (I'm quite certain .pm is ok but not sure about syck)).

The more complete answer is that it is silly to quote n and not m.
http://yaml.org/type/bool.html was a bad stab at boolean from 2005.
After JSON came out the next year, we internally agreed only
treu/false/null should be special plain scalars (although we didn't get
the message out strongly to implementors). I think they should only be
special in [flow, context]. That may happen in the future.

A practical supporting reason to get rid of these is that YAML::Tiny
didn't even do this consistently or roundtrip this:

    $ perl -MYAML::Tiny -e 'print Dump {"n" => "y"}'
    ---
    n: 'y'

    perl -MYAML::Tiny -MData::Dumper -e 'print Dumper Load "n: y\n"'
    $VAR1 = {
                'n' => 'y'
            };