Remove eval around require Scalar::Util
The required version of perl is high enough that Scalar::Util should
always be present.  The eval is only needed because we are also checking
the version.  However, if Scalar::Util is broken for some reason, it's
better to die than silence the error.  If we silence the error, later
attempts to load it will only give "Attepting to reload Scalar/Util.pm
aborted", making diagnostics much harder.