emulate subtest to drop minimum Test::More to 0.88
This commit adds a compatibility shim that provides a subtest-like
function for Test::More 0.88 until subtest was added (circa 0.94, but
not really usable until at least 0.96).

The only major test change is that subtests can't have a plan anymore.
While it was nice to have a plan for test blocks loaded from data files,
it's not really necessary as long as the .t file itself has
done_testing.