Dumping for unknown stuff
This change factors die-ing upon unknown object types (e.g. GLOBs or
blessed objects) into a single method dumper_for_unknown().

The only behavior that is not backwards compatible is that the error
message provided when trying to convert an unknown object at the top
level is no more `Cannot serialize...` but `YAML::Tiny does not
support...` as it was already the case when converting elements inside
arrays and hash values. Performance shouldn't be impacted as well
because the change only affects error conditions.

This allows anyway to sub-class YAML::Tiny (or just re-implement
YAML::Tiny::dumper_for_unknown) and allow customizing the dumping. For
example, it would allow easy support for the plethora of Boolean classes
(JSON::PP::Boolean/Types::Serialiser::Boolean, JSON::XS::Boolean,
Mojo::JSON::_Bool, etc.) with a simple subclassing.

It might be a handy transitional feature while you think about something
for both supporting booleans and supporting generic serialization.
There's anyway a high probability of rejection for this PR, but it's
worth to revive the topics anyway (and it might give you hints/ideas).