Fix for RT #56045, with tests
Fix for correct stripping of leading white space in literal/folded text
blocks - see RT #56045 at https://rt.cpan.org/Public/Bug/Display.html?id=56045

The bugged implementation eats up all the leading whitespace in the
line, not only the amount given by the initial non-empty line as it
should.

The patch includes tests for the fix. It includes a file multiline.tml
that defines a YAML file with string values that have different levels
of internal indentation; the corresponding JSON version includes the
leading spaces in the relevant lines as expected.

For example, in this:

   --- yaml
   - |
      ciao
            a
         tutti
   - |
         quanti
            voi
         amici
   --- json
   ["ciao\n      a\n   tutti\n","quanti\n   voi\namici\n"]

the following happens:

- the indentation for the first array item is related to the word `ciao`, which
has thus no leading space in the JSON rendition. The line immediately following
includes a single `a` character, that is indented with respect to `ciao` by 6
spaces, which are preserved in the JSON string. Same goes for `tutti`, which is
indented by 3 characters with respect to `ciao`.

- the indentation of the second array item is referred to the word `quanti`.
This means that the following word `voi` is further indented by 3 characters
(which reflects in the JSON rendition), while word `amici` is at the same level
as 'quanti` and so it has no leading space in the JSON version.

Call to tests in multiline.tml are added in 30_yaml_spec_tml.t just after the
tests for basic-data.