POC to optionally preserve hash key order
Randomizing key order is fine for Perl's defense against DOS
attacks, but it's pretty sucky for human readable config files.

This is a POC of how to do it with Tie::IxHash.  Using a tied hash
has a big performance hit, but that might be fine for config files.

This commit does not include documentation and has only been minimally
tested.